refactor:Server and Clinet (run func)
